alx_travel_app/requirements.txt
Django 
djangorestframework
django-cors-headers
drf-yasg
celery
rabbitmq